Add functions to read all zones on a single controller
Add new functions to read all of the zones on a single unit.  This
allows querying 6 zones at a time. To query for zones X1 through X6,
these functions can be called with a zone number X0.  Because all of the
zones are sent each with an EOL sequence at the end, add new helper
functions to read a message of a constant size, rather than skipping the
echo and then waiting for the EOL sequence.  An alternate could be to
skip the initial 6 bytes + 5 zone info messages then wait for the EOL on
the last zone if that seems like a better way.